% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_vectors.R
\name{angle_vectors}
\alias{angle_vectors}
\title{Angle Vectors}
\usage{
angle_vectors(vec1, vec2)
}
\arguments{
\item{vec1}{input vector 1}

\item{vec2}{input vector 2}
}
\value{
a named list of the angle between the vectors in degrees and radians
}
\description{
This function finds the angle between two vectors.
}
\examples{
vec1 <- c(1,2,3)
vec2 <- c(2,1,5)
angle_vectors(vec1,vec2)
}
