% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_metrics.R
\name{df_metrics}
\alias{df_metrics}
\title{Data Frame Metrics}
\usage{
df_metrics(df)
}
\arguments{
\item{df}{input data frame from which metrics are calculated}
}
\value{
printed output of the various data frame metrics (sample mean array, covariance matrix, and correlation matrix)
}
\description{
This function outputs relevant metrics for a data frame in column vector and/or matrix form
}
\examples{
df <- Table_1_5
df_metrics(df)
}
