% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_vectors.R
\name{graph_vectors}
\alias{graph_vectors}
\title{Graph Vectors}
\usage{
graph_vectors(vec1, vec2)
}
\arguments{
\item{vec1}{input vector 1}

\item{vec2}{input vector 2}
}
\value{
a figure of the 3D graph of the two vectors
}
\description{
This function graphs two 3D vectors originating from the origin
}
\examples{
vec1 <- c(1,2,3)
vec2 <- c(-1,-2,-3)
graph_vectors(vec1,vec2)
}
