% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Problem_1_4}
\alias{Problem_1_4}
\title{Problem 1.4 Dataset}
\format{
a data frame containing 10 rows and 3 columns:
\itemize{
  \item Sales: Sales in billions of dollars
  \item Profits: Profits in billions of dollars
  \item Assets: Assets in billions of dollars
}
}
\source{
https://www.Forbes.com
}
\usage{
Problem_1_4
}
\description{
A data set containing information regarding the top 10 companies in the world.
}
\keyword{datasets}
