% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_vectors.R
\name{projection_vectors}
\alias{projection_vectors}
\title{Projection Vectors}
\usage{
projection_vectors(vec1, vec2)
}
\arguments{
\item{vec1}{the input vector on which the other vector is projected onto}

\item{vec2}{the input vector which is projected onto the first vector}
}
\value{
the projection of vector 2 onto vector 1
}
\description{
This function serves to project a vector y onto a vector x where the call format (x,y) is used
}
\examples{
vec1 <- c(1,2,3)
vec2 <- c(2,1,4)
projection_vectors(vec1,vec2)
}
