% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subplots_scatter_marginal.R
\name{subplots_scatter_marginal}
\alias{subplots_scatter_marginal}
\title{Subplots Scatter Marginal}
\usage{
subplots_scatter_marginal(df)
}
\arguments{
\item{df}{The data frame input for the figure (must have accurate col names already applied)}
}
\value{
a figure with multiple subplots
}
\description{
This function serves to create a figure layout where subplots are organized into columns and rows. The creation of
    each subplot is handled by the scatter_marginal function.
}
\examples{
df <- Table_1_5
subplots_scatter_marginal(df)
}
