% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table_1_5}
\alias{Table_1_5}
\title{Table 1.5 Dataset}
\format{
a data frame containing 42 rows and 7 columns:
\itemize{
  \item Wind: The wind levels
  \item Solar radiation: The level of solar radiation
  \item CO: The level of carbon monoxide
  \item NO: The level of nitric oxide
  \item NO2: The level of nitrogen dioxide
  \item O3: The level of ozone
  \item HC: The level of hydrocarbons
}
Note: There were no units or scale given with this data
}
\source{
https://www.prenhall.com/statistics
}
\usage{
Table_1_5
}
\description{
A data set containing information regarding the air pollution data in Los Angeles on different days.
}
\keyword{datasets}
