% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec_mean_var_cov.R
\name{vec_mean_var_cov}
\alias{vec_mean_var_cov}
\title{Mean, Variance, Covariance, and Correlation Vectors for Sample Two Vectors}
\usage{
vec_mean_var_cov(x1, x2, x3 = 0)
}
\arguments{
\item{x1}{input vector 1}

\item{x2}{input vector 2}

\item{x3}{option input vector 3}
}
\value{
prints the sample means, sample variances, covariance, and correlation in matrix form
}
\description{
This function calculates common metrics between two vectors including sample mean, sample variance, covariance,
and correlation and returns them in matrix form.
}
\examples{
x1 <- c(1,2,3,4,5,6)
x2 <- c(4,5,3,2,4,6)
vec_mean_var_cov(x1, x2)
}
